[FAB-15279] Split Profile to Network/Context
Removed profile in favor of a network and context combination.  Decoupling the
the network configuration from the context allows for easier toggling between
contexts related to the same network.  Network is a direct reference to a SDK
config yaml.  Context is a container for parameters to be used when working
with the SDK clients.

Added config commands to drive network and context configuration.

Added global flags for changing home directory and current context.

Removed the package scoped functions used to create fabric clients.  Package
scoped functions introduce difficulties with testing.  Accessing clients
through the factory allows for mock replacements.

Created base commands to to reduce redundancy between commands.  The base
commands are embedded in higher level commands.

Change-Id: I7dbafeef1f133cb3c64e9b25e2b274be920b4ab6
Signed-off-by: Brian Buchanan <bpbuch@gmail.com>